+++ README.md
@@ -16,7 +16,7 @@
 
 添加名为  **key**、**SCKEY**、**USER**、**PWD**、**STEP** 的变量，值分别为 **QQ酷推**、**server酱推送key（0关闭）**、 **账号（仅支持手机号）**、**密码**、**步数（0则为1w-2w之间随机 或自定义随机范围[18000-25000]）**
 
-> Settings-->Secrets-->New secret
+> 多账号没测试，自己看着填
